Easier creation of Heavyweight edges in transactions
Added method AddEdge to OTransaction that will handle
connecting the vertices properly during a transaction

Removing accidentally commited *.userprefs and included
a gitignore to prevent adding this file in the future.

The new test cases covers adding a heavyweight edge
and then adding a second edge to check the vectors out_
and in_ hashsets are handled properly.