test fixed for issue #28 and realted to issue #31
This demonstrates a workaround for getting consistent DateTime.
The deserialization of DateTime assumes the server is returning
UTC when database is using JVM default timezone which may be
a local timezone. This should be made more explicit. For now,
this updated test case demonstrates how to get it to work, by
setting the database timezone with "ALTER DATABASE TIMEZONE UTC"